Fix configure using the correct compiler for Android on Mac.
On Mac the folder holding the prebuilt compiler is darwin-x86, not linux-x86.
The extension of PATH to hold the compiler path below ANDROID_NDK_PATH needs to
be present when findtool searches for the compiler, so applying it to the
prefixtool call is not sufficient. OS X has a different behaviour to Linux
which keeps the set value even if it's applied to the prefixtool call only.
Might be caused by OS X coming with an older version of bash.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@29125 a1c6a512-1295-4272-9138-f99709370657