Fix SAPI default speed being way too fast.
The settings currently can't differentiate between default values for different
subitems (i.e. different TTS systems). As a result the default value returned
for SAPI speed was way out of bounds, leading to the upper limit getting used.
Work around this limitation by checking the value for bounds and use a sensible
value in that case.



git-svn-id: svn://svn.rockbox.org/rockbox/trunk@29763 a1c6a512-1295-4272-9138-f99709370657